+++ imouapi/cli.py
@@ -10,20 +10,12 @@ from imouapi.device import ImouDevice, ImouDiscoverService
 from imouapi.device_entity import ImouBinarySensor, ImouSensor, ImouSwitch
 from imouapi.exceptions import ImouException
 
-# configure logging
-loggingconfig = {
-    'level': 'INFO',
-    'format': '%(asctime)s %(levelname)s <%(name)s %(module)s %(funcName)s> %(message)s',
-    'datefmt': '%a, %d %b %Y %H:%M:%S',
-}
-logging.basicConfig(**loggingconfig)  # type: ignore
 
-
-async def discover_devices(app_id: str, app_secret: str) -> None:
+async def discover_devices(app_id: str, app_secret: str, base_url: str, timeout: int) -> None:
     """Discovered devices associated with the account."""
     session = aiohttp.ClientSession()
     try:
-        discover_service = ImouDiscoverService(app_id, app_secret, session)
+        discover_service = ImouDiscoverService(app_id, app_secret, session, base_url, timeout)
         await discover_service.async_connect()
         discovered_devices = await discover_service.async_discover_devices()
         print(f"Discovered {len(discovered_devices)} devices:")
@@ -37,11 +29,11 @@ async def discover_devices(app_id: str, app_secret: str) -> None:
     await session.close()
 
 
-async def get_device(app_id: str, app_secret: str, device_id: str) -> None:
+async def get_device(app_id: str, app_secret: str, device_id: str, base_url: str, timeout: int) -> None:
     """Print out the details of a given device."""
     session = aiohttp.ClientSession()
     try:
-        device = ImouDevice(app_id, app_secret, device_id, session)
+        device = ImouDevice(app_id, app_secret, device_id, session, base_url, timeout)
         await device.async_connect()
         await device.async_initialize()
         await device.async_get_data()
@@ -51,11 +43,13 @@ async def get_device(app_id: str, app_secret: str, device_id: str) -> None:
     await session.close()
 
 
-async def get_sensor(app_id: str, app_secret: str, device_id: str, sensor_name: str) -> None:
+async def get_sensor(
+    app_id: str, app_secret: str, device_id: str, sensor_name: str, base_url: str, timeout: int
+) -> None:
     """Print out the details of a given sensor."""
     session = aiohttp.ClientSession()
     try:
-        device = ImouDevice(app_id, app_secret, device_id, session)
+        device = ImouDevice(app_id, app_secret, device_id, session, base_url, timeout)
         await device.async_connect()
         await device.async_initialize()
         await device.async_get_data()
@@ -70,11 +64,13 @@ async def get_sensor(app_id: str, app_secret: str, device_id: str, sensor_name: 
     await session.close()
 
 
-async def get_binary_sensor(app_id: str, app_secret: str, device_id: str, sensor_name: str) -> None:
+async def get_binary_sensor(
+    app_id: str, app_secret: str, device_id: str, sensor_name: str, base_url: str, timeout: int
+) -> None:
     """Print out the details of a given binary sensor."""
     session = aiohttp.ClientSession()
     try:
-        device = ImouDevice(app_id, app_secret, device_id, session)
+        device = ImouDevice(app_id, app_secret, device_id, session, base_url, timeout)
         await device.async_connect()
         await device.async_initialize()
         await device.async_get_data()
@@ -90,11 +86,13 @@ async def get_binary_sensor(app_id: str, app_secret: str, device_id: str, sensor
     await session.close()
 
 
-async def get_switch(app_id: str, app_secret: str, device_id: str, sensor_name: str) -> None:
+async def get_switch(
+    app_id: str, app_secret: str, device_id: str, sensor_name: str, base_url: str, timeout: int
+) -> None:
     """Print out the details of a given switch."""
     session = aiohttp.ClientSession()
     try:
-        device = ImouDevice(app_id, app_secret, device_id, session)
+        device = ImouDevice(app_id, app_secret, device_id, session, base_url, timeout)
         await device.async_connect()
         await device.async_initialize()
         await device.async_get_data()
@@ -110,11 +108,13 @@ async def get_switch(app_id: str, app_secret: str, device_id: str, sensor_name: 
     await session.close()
 
 
-async def set_switch(app_id: str, app_secret: str, device_id: str, sensor_name: str, value: str) -> None:
+async def set_switch(
+    app_id: str, app_secret: str, device_id: str, sensor_name: str, value: str, base_url: str, timeout: int
+) -> None:
     """Print out the details of a given switch."""
     session = aiohttp.ClientSession()
     try:
-        device = ImouDevice(app_id, app_secret, device_id, session)
+        device = ImouDevice(app_id, app_secret, device_id, session, base_url, timeout)
         await device.async_connect()
         await device.async_initialize()
         await device.async_get_data()
@@ -127,7 +127,7 @@ async def set_switch(app_id: str, app_secret: str, device_id: str, sensor_name: 
                     await sensor.async_turn_off()
                 elif value == "TOGGLE":
                     await sensor.async_toggle()
-                await get_switch(app_id, app_secret, device_id, sensor_name)
+                await get_switch(app_id, app_secret, device_id, sensor_name, base_url, timeout)
                 break
     except ImouException as exception:
         print(exception.to_string())
@@ -141,8 +141,16 @@ class ImouCli:
         """Initialize."""
         self.app_id = None
         self.app_secret = None
+        self.base_url = None
+        self.timeout = None
+        self.logging = "INFO"
         self.device_id = None
         self.command = None
+        self.loggingconfig = {
+            'level': 'INFO',
+            'format': '%(asctime)s %(levelname)s <%(name)s %(module)s %(funcName)s> %(message)s',
+            'datefmt': '%a, %d %b %Y %H:%M:%S',
+        }
         self.args = []
 
     def parse_command_line(self):
@@ -162,6 +170,20 @@ class ImouCli:
                 self.app_secret = sys.argv[i + 1]
                 skip_next = True
                 continue
+            if arg == "--base-url":
+                self.base_url = sys.argv[i + 1]
+                skip_next = True
+                continue
+            if arg == "--timeout":
+                self.timeout = sys.argv[i + 1]
+                skip_next = True
+                continue
+            if arg == "--logging":
+                self.logging = sys.argv[i + 1]
+                self.loggingconfig["level"] = self.logging.upper()
+                logging.basicConfig(**self.loggingconfig)  # type: ignore
+                skip_next = True
+                continue
             if self.command is None:
                 self.command = arg
                 continue
@@ -175,30 +197,48 @@ class ImouCli:
             self.print_usage()
             sys.exit(1)
         if self.command == "discover":
-            asyncio.run(discover_devices(self.app_id, self.app_secret))
+            asyncio.run(discover_devices(self.app_id, self.app_secret, self.base_url, self.timeout))
         elif self.command == "get_device":
             if len(self.args) == 1:
-                asyncio.run(get_device(self.app_id, self.app_secret, self.args[0]))
+                asyncio.run(get_device(self.app_id, self.app_secret, self.args[0], self.base_url, self.timeout))
             else:
                 print("ERROR: provide device id")
         elif self.command == "get_sensor":
             if len(self.args) == 2:
-                asyncio.run(get_sensor(self.app_id, self.app_secret, self.args[0], self.args[1]))
+                asyncio.run(
+                    get_sensor(self.app_id, self.app_secret, self.args[0], self.args[1], self.base_url, self.timeout)
+                )
             else:
                 print("ERROR: provide device_id and sensor_name")
         elif self.command == "get_binary_sensor":
             if len(self.args) == 2:
-                asyncio.run(get_binary_sensor(self.app_id, self.app_secret, self.args[0], self.args[1]))
+                asyncio.run(
+                    get_binary_sensor(
+                        self.app_id, self.app_secret, self.args[0], self.args[1], self.base_url, self.timeout
+                    )
+                )
             else:
                 print("ERROR: provide device_id and sensor_name")
         elif self.command == "get_switch":
             if len(self.args) == 2:
-                asyncio.run(get_switch(self.app_id, self.app_secret, self.args[0], self.args[1]))
+                asyncio.run(
+                    get_switch(self.app_id, self.app_secret, self.args[0], self.args[1], self.base_url, self.timeout)
+                )
             else:
                 print("ERROR: provide device_id and sensor_name")
         elif self.command == "set_switch":
             if len(self.args) == 3:
-                asyncio.run(set_switch(self.app_id, self.app_secret, self.args[0], self.args[1], self.args[2]))
+                asyncio.run(
+                    set_switch(
+                        self.app_id,
+                        self.app_secret,
+                        self.args[0],
+                        self.args[1],
+                        self.args[2],
+                        self.base_url,
+                        self.timeout,
+                    )
+                )
             else:
                 print("ERROR: provide device_id and sensor_name")
         else:
@@ -209,9 +249,12 @@ class ImouCli:
         print("imouapi cli")
         print("Usage: python -m imouapi.cli [OPTIONS] COMMAND <ARGUMENTS>")
         print("")
-        print("Options (mandatory):")
-        print("  --app-id <app_id>                                      Imou Cloud App ID")
-        print("  --app-secret <app_secret>                              Imou Cloud App Secret")
+        print("Options:")
+        print("  --app-id <app_id>                                      Imou Cloud App ID (mandatory)")
+        print("  --app-secret <app_secret>                              Imou Cloud App Secret (mandatory)")
+        print("  --logging <info|debug>                                 The logging level")
+        print("  --base-url <base_url>                                  Set a custom base url for the API")
+        print("  --timeout <timeout>                                    Set a custom timeout for API calls")
         print("")
         print("Commmands:")
         print("  discover                                               Discover registered devices")
@@ -219,7 +262,7 @@ class ImouCli:
         print("  get_sensor <device_id> <sensor_name>                   Get the state of a sensor")
         print("  get_binary_sensor <device_id> <sensor_name>            Get the state of a binary sensor")
         print("  get_switch <device_id> <sensor_name>                   Get the state of a switch")
-        print("  set_switch <device_id> <sensor_name> [on|off|toggle]  Set the state of a switch")
+        print("  set_switch <device_id> <sensor_name> [on|off|toggle]   Set the state of a switch")
 
 
 # create an instance of the cli