+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.3] (2022-09-27)
+### Changed
+- API base URL and API timeout are not optional parameters and can be provided by the user
+- Updated CLI, added logging level
+- Updated documentation
+
 ## [0.1.2] (2022-09-27)
 ### Changed
 - Re-organized the file structure